refactor(argocd): move the rollout report into argocd and align state backends
Relocate the rollout failure-report renderer, the image-name and retry-math
helpers, and the desired-image walk from internal/models and internal/helpers
into internal/argocd next to their only callers, leaving models as data types.
Drop the unused models.LockdownSchedule.

Add a repository contract suite that runs against both state backends. It
exposed PostgresState.AddTask returning created in milliseconds and leaving
updated unset on the task handed to the start notification; both are now Unix
seconds. SetTaskStatus returns ErrTaskNotFound in both backends.
