+++ internal/argocd/image_validation_test.go
@@ -49,11 +49,11 @@ func TestShouldValidateDesiredImages(t *testing.T) {
 		status   string
 		expected bool
 	}{
-		{"settledWithoutImage", "Synced", "Healthy", models.ArgoRolloutAppNotAvailable, true},
-		{"stillSyncing", "OutOfSync", "Healthy", models.ArgoRolloutAppNotAvailable, false},
-		{"stillProgressing", "Synced", "Progressing", models.ArgoRolloutAppNotAvailable, false},
-		{"imageAlreadyThere", "Synced", "Healthy", models.ArgoRolloutAppSuccess, false},
-		{"degraded", "Synced", "Degraded", models.ArgoRolloutAppDegraded, false},
+		{"settledWithoutImage", "Synced", "Healthy", ArgoRolloutAppNotAvailable, true},
+		{"stillSyncing", "OutOfSync", "Healthy", ArgoRolloutAppNotAvailable, false},
+		{"stillProgressing", "Synced", "Progressing", ArgoRolloutAppNotAvailable, false},
+		{"imageAlreadyThere", "Synced", "Healthy", ArgoRolloutAppSuccess, false},
+		{"degraded", "Synced", "Degraded", ArgoRolloutAppDegraded, false},
 	}
 
 	for _, test := range tests {
