+++ internal/argocd/testdata/out-of-sync-multi-source.json
