+++ internal/models/managed_resources.go
@@ -1,12 +1,5 @@
 package models
 
-import (
-	"encoding/json"
-	"slices"
-
-	"github.com/shini4i/argo-watcher/internal/helpers"
-)
-
 // ManagedResources is the response of ArgoCD's
 // /api/v1/applications/{name}/managed-resources endpoint: the desired state of every
 // resource the application manages. It is the only reliable answer to "does this image
@@ -23,63 +16,4 @@ type ManagedResources struct {
 // exists only in the cluster.
 type ManagedResource struct {
 	TargetState string `json:"targetState"`
-}
-
-// DesiredImageNames returns the sorted, de-duplicated repository names (tags and
-// digests stripped) of every container image declared in the application's desired
-// state. An unreadable item is skipped rather than aborting the walk: ArgoCD marshals
-// every target state, so an item that fails to decode carries nothing to read anyway.
-func (resources *ManagedResources) DesiredImageNames() []string {
-	if resources == nil {
-		return nil
-	}
-
-	found := make(map[string]struct{})
-
-	for index := range resources.Items {
-		targetState := resources.Items[index].TargetState
-		if targetState == "" {
-			continue
-		}
-
-		var manifest any
-		if err := json.Unmarshal([]byte(targetState), &manifest); err != nil {
-			continue
-		}
-
-		collectImages(manifest, found)
-	}
-
-	names := make([]string, 0, len(found))
-	for name := range found {
-		names = append(names, name)
-	}
-	slices.Sort(names)
-
-	return names
-}
-
-// collectImages records the repository name of every string stored under an "image"
-// key, at any depth. Matching on the key alone rather than on known pod-template paths
-// keeps Deployments, CronJobs, Rollouts and image-bearing CRDs on one code path.
-// A stray match widens the accepted set, but it also makes the set non-empty, which is
-// what lets the caller conclude at all — see DesiredImageNames' callers for how an empty
-// set is treated.
-func collectImages(node any, found map[string]struct{}) {
-	switch value := node.(type) {
-	case map[string]any:
-		for key, child := range value {
-			if image, ok := child.(string); ok {
-				if key == "image" && image != "" {
-					found[helpers.ImageName(image)] = struct{}{}
-				}
-				continue
-			}
-			collectImages(child, found)
-		}
-	case []any:
-		for _, item := range value {
-			collectImages(item, found)
-		}
-	}
 }
